fix(si-web): fix qualifications init state (#397)
When we return from updating the entity, we included the qualifications.
But we include the *old* qualifications, which means we actually have a
race with the websocket updating the state. This stops us from using the
value we reply with, and instead just letting the web socket do its
business.