fix(sdf): always qualify production
We were qualifying 'baseline' sometimes, which means we were frequently
not having all the data we needed to qualify an object. This PR makes it
so that, no matter what, we always grab the production system and use
its ID (unless we're given a specific one).

This will fix it for now, but it's wrong long term. We should always
send a system with a qualification request, or we should qualify every
system.