feat(*): add v1 of discovery (#401)
This PR adds discovery to SI. It allows you to specify that a given
implementation is `discoverableFrom` another. For example, an 'awsEks'
implementation is discoverableFrom `aws`.

When you view the new discovery sub-panel on `aws`, you will see a list
of entity types that are discoverable. Hit the refresh button, and we
will discover all the existing implementations of that kind within your
account. This must be done *outside of a change set or edit session*.

You can then go to, say, a `kubernetesCluster` node that has importable
implementations. While still outside of a change set or edit session,
you can import the implementation into your current schematic. This
happens outside of a change set because, unfortunately, there is no
other way to make sure the relevant objects all exist in head (which is
a requirement for making the whole thing work).

The implementation and its nodes will appear on the graph already
connected, but there will be no connection between the implementation
and the concept. You do that yourself in a new edit session.