feat(*): Add discovery of k8s objects (#402)
This adds a v1 of k8s objects. It causes some errors in the dashboards,
because we were being pretty silly about how we decide what counts as an
update. You can always hard refresh to make it better.

Otherwise, it does successfully discover the 'play-2048' service from
the cluster. Don't try anything else, as I make no guarantees. :)

You should be able to deploy the service as well.