chore(si-sdf): add initial telemetry infrastructure
Hey, look at that, telemetry!

This is a rather large/churn-y code change but mostly does make sense to
see it as a unit. The goal here is to re-introduce a way for us to
follow the flow of a user task or system operation as it propagates
through the various services. As before, we're still ultimately
targeting HoneyComb as our backend for tracing and secondarily for
logging and events (more to follow on those later). We're using tooling
and specification work from the [OpenTelemetry
project](https://opentelemetry.io/) both as a way of architecting some
observability into our system, but also as a way of normalizing the
traffic, choosing protocols, etc.

OpenTelemetry Collector
-----------------------

First up is a slimmed down OpenTelemetry Collector image which is used
in development to collect tracing from sdf and veritech. The idea here
is to leave as much deployment environment detail to...the environment
and out of the application. As such the OpenTelemetry SDK
implementations will emit to a locally running collector by default.
Let's use that, shall we?

The image can be managed by changing to the `components/otelcol` (that
is: otel == OpenTelemetry/col == Collector) directory. There are 2
environment variables each developer will need to set:

- `HONEYCOMB_TOKEN`: a HoneyComb API token that grants permission to
  write to a dataset--that's the only permission we would need to grant
  to those keys.
- `HONEYCOMB_DATASET`: the target HoneyComb dataset name, typically
  `dev-$USER` by convention for each developer.

These variables only need to be set when the container is created (that
is, during the initial `docker run` step), but a reasonable approach is
to set these variables in an `.envrc` file in the project root which is
Git-ignored by default. At least, that's yours truly's approach.

 ### Running a Fresh Collector

```sh
cd ./components/otelcol
make run
```

 ### Stopping a Container

```sh
make stop
```

 ### Tailing the Container Logs

```sh
make tail
```

 ### Clean Up the Container

```sh
make clean-container
```

 ### Clean Up the Container and Image

```sh
make clean
```

It should be noted that in Kubernetes production we'd likely run a
daemonset of collectors (that is, one per node) which forward to a small
deployed service of the final collectors which would send to HoneyComb.
As such, there isn't much use for our little container image in
production--better solutions there.

Rust PostgreSQL Instrumentation
-------------------------------

A large part of this work was instrumenting a substantial part of the
database client code which is used primarily by `si-model` and `si-sdf`.
This was accomplished by, in effect, wrapping the underlying
tokio-postgres and deadpool-postgres libraries with our own layer that
can instrument each useful function. This allows us to track many nice
things such as:

- recording connection details, especially useful once in production
  against potentially multiple/different AWS instances
- recording the incoming queries, noting that we don't currently track
  the params yet due to privacy concerns
- recording the number of resulting rows returned by queries where
  appropriate
- linking relevant queries to the span representing a transaction call,
  allowing us to see all queries performed "under" that transaction
- recording if a transaction was ultimately committed or rolled back
  which will be useful to catch future bugs

Rust Warp/API Initial Instrumentation
-------------------------------------

This work isn't complete or fully comprehensive, but rather a start with
some endpoints having more detail and all having some degree of
instrumentation. For the web server framework Warp, we're injecting some
middleware which generates a new span for each request (at the moment
generating a fresh root trace). Each of these request spans are
recording information such as URL path, remote client address, HTTP
method, etc. Future work will add other information such as the final
HTTP status code, etc.

The authentication code was also updated to be handled in the filter
pipeline before delegating to the handler functions. This means that we
can enforce authentication right up front and not have to additionally
remember to do this in our near-business logic code. This also means
that all the authenticated endpoints' handlers are provided with an
`SiClaims` containing the user id and billing account id which will help
when adding richer spans to cover all those handlers off.

Rust Veritech Client Instrumentation
------------------------------------

In order to better track the hand off from sdf to veritech, the veritech
client (that is `veritech.rs` in `si-sdf`) has been fully instrumented
to track the HTTP client request/response loop. This includes the more
straight forward synchronous HTTP GET/POST calls and the spawned
WebSocket calls. OpenTelemetry context propagation is used to ensure
that the current span in sdf is injected into HTTP headers and then
extracted on other side to become the parent span there.

There is likely still work to do on the WebSocket side as there is a
little bi-directionality but it's enough for the moment to track who is
calling who.

TypeScript Veritech Server Instrumentation
------------------------------------------

Similarily to the Rust side, Veritech is also instrumented from multiple
libraries from the OpenTelemetry ecosystem. Our web server framework Koa
has instrumentation/middleware installed which will spawn request spans
for all requests that have incoming span context data. The
JavaScript/Node auto-instrumentation seems pretty good as it comes out
of the box with good resource detectors, injecting rich information
about the sdk/version of the libraries.

As before, there's still likely some future work here to get more
granularity but it won't be a herculean task. Note that as far as I can
tell in this ecosystem, there's much more need for explicit passing of
parent spans down into functions which need to manually wire up the new
children. This may be in part due to our use of the Tokio tracing
ecosystem on the Rust side and that being so darn good. Then again, I
may have missed some JavaScript injection insanity too which would make
child span creation a little more magical but ergonomic.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>