chore: rebuild Docker images for key services
This is the first iteration of a rebuild of Docker images for the
following services:

- si-sdf
- si-veritech
- si-web-app

They are not fully complete but with a couple of caveats, are functional
together. Crucially, the demo flow works aside from the dockerImage
qualification check as mentioned below.

Currently, they are built to work with Docker Compose, with an example
in `support/dev`. To build all the images, you can:

```sh
cd support/dev
./demo-build.sh
```

which invokes `docker-compose` in parallel build mode. Some Honeycomb
credentials are required and the script should fail if they aren't found
in the current shell environment.

To launch the setup run: `docker-compose up` in the `support/dev`
directory.

All the relevant ports are exposed out to the host system, meaning that
all the same local ports are served up--this time via the container
services. So you can, in theory develop on `si-web-app` as normal and
have the rest of the stack running with Docker Compose

Caveats
-------

Currently there are some known issues to address:

- Which Veritech's image has an installation of Docker, it's not yet
  wired and capable of using the outer engine for dockerImage
  qualifications
- SDF is currently using a local PostgreSQL container and needs to have
  Aurora credentials passed through to the SDF container on start
- The si-web-app container needs a slightly customized nginx
  configuration to use the `try_file` directive to help with some client
  refreshing/routing back to `/index.html`

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>