chore: initially deploy si on k8s
This is a first iteration of a Kubernetes deployment of SI. There were 2
initial targets:

- minikube
- AWS EKS

Both deployments make use of the ingress-nginx controller (`minikube
addons enable ingress`) and in the EKS case, the ingress controller
service is fronted with an auto-deployed NLB (the service type is
`LoadBalancer` with metadata annotations).

The Kubernetes deployments can use Kustomize under the hood with a
suitably setup `kubectl`:

```sh
cd support/dev/k8s
kubectl apply --namespace si-dev -k default/
```

Note that this assumes the namespace is created and this overlay will
deploy a pretty stateless postgres instance (it is not currently a
statefulset).

To deploy an overlay which uses the deployed Aurora database instance:

```sh
cd support/dev/k8s
kubectl apply --namespace si-dev -k eks-aurora/
```

There are also several secret/crediential files which need to be filled
out for the secrets to be generated and deployed, namely filling in:

- support/dev/k8s/base/dockerconfig.json (see dockerconfig.json.sample)
- support/dev/k8s/base/honeycomb.env (see honeycomb.env.sample)
- support/dev/k8s/base/web-htpasswd

For the default overlay, additionally:

- support/dev/k8s/default/postgres.env (see postgres.env.sample)

and for the eks-aurora overlay:

- support/dev/k8s/eks-aurora/postgres.env (see postgres.env.sample)

Caveats
-------

Currently there is a primary issue in the EKS deployment:

- It appears as though the NLB is eventually convinced that the backends
  (that would be the nginx ingress service) in unhealthy and stops
  routing traffic back to it. This causes the inbound traffic to block
  at the NLB and not get further routed. Additionally, active web socket
  connections are severed and don't receive further updates. This takes
  on average a little over 5 minutes, which is a very fishy amount of
  time and reeks of a timeout period. Minikube's great though!

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>