bug(si-entity): YAML numbers now schema driven (#440)
Previously, we scanned the entire property tree for anything that looked
like a number - and if it did, we turned it into a number when we
serialized. This broke the contract with k8s, because sometimes it
requires that everything be a string (for example, in an annotation).

This replaces that brute-force method with one that uses the ops on the
entity to determine if they are a number according to the schema, and if
they are, *then* replace them with the numeric representation rather
than a string. Otherwise they remain a string.

It also cleans up failing test from previously.