bug(si-web): allow textareas in map values
This allows you to set the widget for a valueProperty of a map in the
schema to be a 'textArea'. If you do, you will get a one line sized text
area, that acceps the raw input - suitable for pasting, writing YAML
directly, or whatever your heart desires.

It also displays the output correctly when viewing it not in edit mode.

This change updates the k8sConfigMap and k8sSecret to have their `data`
maps have their value set to textArea.

Caveats:

* I would not say this is a great user experience for editing big inline
  data like this - we probably should think about how using Monaco might
  work even better. Maybe a way to switch from a textArea to monaco?
  Could be neato.