bug(si-entity): yaml number replacement improvements (#447)
This should resolve the bug where we do not correctly translate between
numbers in the schema and numbers in YAML. It uses a hybrid approach,
where the property object itself is walked, and each path we encounter
becomes a fake op, which we use to look up the correct prop from the
schema. If that prop declares it should be a number, then we swap the
value to a number.

This PR also adds more test coverage for both finding properties (which
needed to be re-written for this to work) and the eventual results of
YAML numbers.