build(image): prefer `gdate` when computing build time & version info
We're preferring, or rather first checking for a `gdate` binary on
`$PATH` so that the release process can be run on macOS as well as
Linux systems, which is the current supported default.

On macOS, when using Homebrew a developer can install the `coreutils`
package which will install GNU versions of coreutils programs on
`$PATH`, with all programs prefixed with a `g`. In this way, our release
process should flow just as naturally on macOS as on Linux.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>