build(image): optionally push `:latest` tag on publish
Note that this only affects a publish task, and not a release task.

When using the `build-image.sh` scripts directly, a publish with a push
to latest is invoked with: `build-image.sh --push --latest`.

When using the `make publish` task, a publish with a push to latest is
invoked with: `make publish PUBLISH_ARGS=--latest`.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>