feat(*): allow linking to existing nodes
Enable linking existing nodes into the current context. You can link to
an existing node via the Link menu. If the menu is enabled, it will
follow the same path as the add menu, but the individual nodes of that
type in the workspace (that are not in the current view) will be
present. You can then add them to the current schematic.

If the node you are linking to has incoming edges, we are *not*
currently pulling those nodes in along for the ride. So if you want to
see all of the configuration, you need to add all those nodes as well
via the link menu. We expect that another PR will flesh out that side of
the work - essentially make it so that when an edge is created for a
given entity, we ensure the inbound entities are also in the same
contexts. But tomorrows problem. :)