fix(si-veritech): map all k8sDeployment ports into k8sService (#458)
This change fixes an issue in the `k8sService`'s infer properties
intelligence. Rather than build and appending a object of port entries,
the code was pushing single objects of with only a `port` or `protocol`
(but not both, together). In short, this author has no idea how we were
getting anything of value to be derived. Likely "bad data" was thrown
out at a certain stage and the last bit of remaining "good data" was the
last entry with a `port` key in it. But now it's all here, cool!

Note that likely replaying this intelligence could lead to duplicate
entries or a different order. There may be more work to do here, but
it's also working within the strengths and limitations of where this
code is at. A little "Demoer Beware" should go far.

References: Create a port entry for each exposed port [ch1475]

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>