feat(model,registry,veritech): connect dockerHubCredential to k8sSecret
This change allows a single `dockerHubCredential` to infer the type and
data payload of a `k8sSecret`. Currently the protected SI secret is
decrypted and serialized into the `k8sSecret` body and is therefore no
longer protected, so take note of that, okay?

Prior to this change decrypted secrets were not being passed into
`inferProperties` calls (as they should not leak private credentials),
but for the moment this unlocks the functionality before we close this
loophole by other means. As a result, this is a churny change in
Veritech and SDF.

To support the intelligence functions a new helper
`setPropertyFromEntitySecret` is added which has a mandatory callback
called `transform` which yields the decrypted `message` payload as a JS
object and returns a return type of `string | number | boolean` so we
can set the result as a property value ultimately.

And finally, the existing `setPropertyFromEntity` helper gains an
optional `transform` callback, yielding the value plucked from the
context's entity and the expects a return type of `string | number |
boolean` for a final `OpSet`.

References: ability to connect a dockerHubCredential [ch1484]

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>