feat(*): add si-inference and si-cyclone
This PR adds two new components:

1. si-inference

This is mostly a placeholder for the types needed for a low level
inference call, because it will be depended on both in the browser and
in veritech.

2. si-cyclone

This is a new package, named after Robotech's Veritech Cyclone, which
was a small powered armor that converted to a motorcycle. It's job is to
execute inference functions inside a secure javascript runtime, using
vm2. It will eventually integrate with vi-veritech for execution.

It also further extends the schema, adds some functionality to
si-entity, and adds lots of tests.

It is currently un-integrated with the rest of the system. The best way
to see things in action is to run `npm run test` in si-cyclone, and
watch the test suite do its thing.