build(deps): update Rust dependencies (#469)
This update leaves out the upgrade to clap 3.0.0-beta.4 as there are API
breaking changes that will take more than a few minutes to deal with. As
before, the `tracing` ecosystem crates are patched to use a Git
checkout, updated to the current commit on the `v0.1.x` branch, still
waiting on a new release of `tracing-subscriber` to revert to mainline
releases. The `tungstenite` crate is also untouched as
`tokio-tungstenite` hasn't released a new version as of today.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>