feat: add array edit partial support for custom per-entry edit fields
This feature is to support the ability for a user to select a custom
"view" or subset of an object when adding Kubernetes environment entries
in a Deployment/StatefulSet/DaemonSet. The Kubernetes specification for
these environment variable entries to be one of 5 possible kinds (where
each kind refers to the source of the value):

* manually setting the `value` field with a value
* pulling the value from a deployed `ConfigMap`, at a specific key
* pulling the value from a field of the pod spec itself, for example
  from the `metadata.name`, `metadata.annotations[KEY]`, etc (see the
  spec for more details)
* pulling the value from the resources of the container, such as
  `limits.cpu`, `requests.memory`, etc.
* pulling the value a deployed `SecretKey`, at a specific key

As each of these value sources are mutually exclusive--that is, it is
semantically invalid to populate both the `value` field and the
`valuFrom.configMapRef` object for example--it would be beneficial to
only display one of these potential "views" at each array entry. That is
what this change allows for.

The registry schema has been extended in a limited way, by adding an
optional `editPartials` entry on `PropObject`s (and by extention
`PropObjectItem`s), which allow a schema author to describe zero or more
edit partials which have a name and a set of property paths that are
paths relative to the targeted object. These edit partials can be
grouped under zero or more categories which allow for tree-like
organization, similar to the add menu categories. Currently, edit
partials are only used in edit mode in the attribute viewer and *not* in
read-only mode. If the type of partial currently being used wants to be
changed by an end user, they will have to delete the old entry and add a
new entry with the different partial type. This shouldn't be a problem
in practice as the Kubernetes spec describes environment variables as an
array of entries (as opposed to a map/hash/directory), so presumably the
last entry "wins" over old ones, even if there are 2 or more entries
corresponding to the same name.

References: improve the ux of adding container environment [ch1495]

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>