fix(si-veritech): fail k8s obj apply when no cloud providers are found (#473)
This change slightly alters the shared behavior for all Kubernetes
default `apply` action intelligence. In the intelligence, we currently
look for an AWS EKS Cluster and if one if found, we run `kubectl apply`.
And we do the same check and execution for an Azure AKS Cluster. If
*neither* are found, then there was no commands executed, no output
emitted, and no failure propagated.

With this change, if no command executions are run, then an error line
is emitted reporting that the intelligence expected to execute
something, but nothing resulted. Finally an exception is thrown which
causes the action to result in a *failure* (rather than the default
*success*). While this might not be a traditional systems failure, from
an SI user's point of view, given a correctly connected graph, something
*should* happen. So for now, we'll consider this a failure and track
this as we go forward.

References: action finishes with success but it never [ch1541]

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>