feat(cyclone): support optionally limiting requests
This change introduces 2 new CLI options:

* `--oneshot`
* `--limit-requests <N>`

both of which will limit the number of requests to service before
issuing a graceful shutdown of the server and process. The `--oneshot`
flag is an alias for `--limit-requests 1` and is likely to be the normal
operating mode.

Additionally the following internal changes are made to cyclone:

* A formal graceful shutdown mechanism was introduced which allows the
  axum server to drain connections before terminating. There are
  currently 2 paths to signal a graceful shutdown: 1) issuing a
  `SIGTERM` signal to the process and 2) using request limiting.
* A refactoring of the `Server` to use generics rather than an internal
  enum to ultimately simplify and unify the HTTP and UDS server flavors.
  This is in anticipating of adding an HTTPS flavor very soon.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>