refactor(veritech): update component shape, in prep for pooling
This is largely a setting the table change to `si-veritech` to support
active pooling of cyclone workers. Some of the changes also updated
`si-cyclone` to make it a better client library to consume--largely this
was adding a `CycloneClient` trait with the main methods to better pass
around different generic types of the built clients. So far, this
avoided the extra runtime of needing a trait object, which while it
wasn't the goal, was a pretty nice side-effect.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>