chore(deadpool-cyclone): restructure for additional impls & document
This change changes the crate setup to make adding additional Cyclone
Instance impls easier. It also documents the crate for
use/consumption/maintenance by a broader team, outside of yours truly
author ;)

P.S. Also, this is the first member crate that was updated to the Rust
2021 edition and therefore has a minimum Rust version of 1.56.0--which
is now enforced by Cargo via the `rust-version` entry in `Cargo.toml`.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>