chore: implement tunable/config'able db pooling (#525)
This change is fairly cross cutting, but is in support of configurable
and tunable database pooling--primarily for SDF at the moment--from
configuration (i.e. environment, configuration file, CLI arguments), to
server boot, to server setup, to available resources in each route
handler.

Here's a little color on the details:

Service Configuration
---------------------

There are now 4 sources of configuration input for SDF, Veritech, and
any other server that may require it (note that Cyclone is a specialized
case and does *not* have configuration at this time).

1. The config file defaults. For SDF (and Veritech) there is a
   `ConfigFile` type which supports `Deserialize`, `Serialize` and most
   importantly, `Default`. The defaults will be taken from this type as
   used as the base level for all other sources as described below.
2. A configuration file. At the moment, only support for TOML formats
   are enabled but we can add JSON and YAML if desired. The name of the
   configuration file matches the name of the binary, so in the case of
   the `sdf` binary, the configuration file is `sdf.toml` if present.
   There are 3 locations that will be checked on the file system:

   1. The current directory of the running process. This could be the
      directory of the binary, but rather it always corresponds to
      `std::env::current_dir()` from the Rust standard library. So the
      target would be the same as `$(pwd)/sdf.toml`.
   2. A local user configuration. On Linux this follows the XDG
      specification, so SDF's configuration would be checked for in the
      equivalent of `$HOME/.config/sdf/sdf.toml`.
   3. A system-wide configuration, either under `/usr/local/etc` or
      `/etc`. That is, `/usr/local/etc/sdf/sdf.toml` would match and
      then `/etc/sdf/sdf.toml` would match next.

   An environment variable of the form `SI_${NAME}_CONFIG` is first
   checked, and if set, this value will be used before looking on the
   file system as described above. In SDF's case, the environment
   variable configuration location would be `SI_SDF_CONFIG`.

   Note that the contents of this configuration file can be sparse, that
   is not every value/key must be set. Anything missing will inherit
   either defaults, or values set with hiher precedent sources, as
   described below.
3. Environment variable to override a terminal/scalar value in a
   configuration file. The environment variable prefix is of the form
   `SI_${NAME}` with double underscores separating the hierarchical
   sections of configuration. For example, setting the `db.name` for
   SDF, requires an environment variable like this:
   `SI_SDF__DB__NAME=myappdbuser`.
4. CLI arguments. Not necessarily every configuration value is settable
   from the CLI--this is by design and intentionally so. CLI argument
   parsing continues to use the `clap` v3 crate implementation with the
   derive macro to get nice parsable types. Rather, SDF's binary program
   (which owns the CLI parsing type) derives a conversion from the CLI
   args into the `ConfigFile` type, by using a helper function to set
   the equivalent values into a HashMap like structure for later merging
   with the other sources.

All of these sources and layers are loaded and merged via the
`config-file` crate which uses many tricks and other crates such as
`config`, `pathdiff`, as well as a custom Serde serializer. In short,
it's painful how bad unifying configuration is in the Rust ecosystem,
but I believe we have a kick ass implemention to use, and it's all
tucked away in a crate.

The final SI-specific magic continues to exist in the `si-settings`
crate but rather is driven from 2 new traits which set up the `Config`
and `ConfigFile` behavior to allow this all to work in the end
applications.

SDF Service Booting
-------------------

Once we have our runtime configuration figured out, it's time to boot up
the server components. First up is establishing our PostgreSQL
connection pool. The `PgPoolConfig` type has migrated from
`si-settings` into `si-data` to house the configuration with the
implemention--any consumers of the `PgPoolConfig` type would necessarily
as want to consume the `PgPool` behavior. Not much has changed here,
except in one important detail:

When establishing a connection pool, a task is spawned which attempts to
get a connection from the pool and perform a simple validating query
(`SELECT 1`) as a test of being able to round-trip to the database.
However, any failures in this test will not crash or stop the server
process anymore. Any issues will be logged at the `WARN` level however
so that later monitoring or operator inspection can happen. This change
was made to acknowledge that the database may or may not be reachable
when the service starts but it may appear at some point in the near
future. The `PgPool` (using the `deadpool` crate as implementation) has
mechanisms for keeping a pool healthy and recycling old/disconnected
clients, so we should fully embrace that.

The NATS client gets a similar treatment as the `PgPool` above, except
there is no pool of client connections--only the one which gets cloned
and shared among multiple concurrent parts of the system. Future work
will augment this client, make it more tunable, and likely perform a
similar test connection task trick.

SDF Database Migration
----------------------

Before the web server component has started and after we have
established a PG pool, we perform any database migrations as before.
This change however introduces 3 database migration modes/strategies
which are configurable via the `--migration-mode` CLI option, the
`SI_SDF__MIGRATION_MODE` environment variable, or a `migration_mode`
entry in a TOML configuration file. The 3 modes are:

1. `run`: runs the migration, cool!
2. `skip`: skips over the migration and continues to boot. When we run
   SDF as a service cluster it will likely be adventageous to run all
   normal traffic-serving instance in the `skip` mode
3. `runAndQuit`: runs the migration, then immediately ends the process.
   In production we can deploy a single instance (maybe in a canary
   deployment, something like that) in the `runAndQuit` mode, then
   proceed to launch the rest of the traffic-serving instances in
   `skip` mode.

SDF Database and NATS Axum Extractors
-------------------------------------

To help provide a more ergonomic experience when writing handlers,
several custom extractors are provided which can be used as parameters
in the Rust handler functions. You can check out
`lib/sdf/src/server/handlers.rs` for the `demo` handler for examples :)

Maybe More?
-----------

There's a lot here, next time smaller chunks.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>