chore(web): fix websocket restart, refactor events and services (#537)
This makes the web app correctly deal with login/logout state, and ensure that the websocket is refreshed correctly when HMR restarts the app.

It also moves the 'api/sdf/service/*' code to be top level. Services bind together the SDF API with the applications state (stored in observables, or xstate, or whatever we want).

It also moves event dispatch to a service, which makes it easy to both process incoming events and dispatch local events.