feat(web): global error service (#538)
This introduces a GlobalErrorService, which handles updating the state for the global error display. (that gets mounted in the Home page.)

It also fixes a small bug in the websocket being created when there is no current authentication token (which isn't a thing), and moves a button around in the change set.