fix(si-data): fix NATS subscription stream impl
This change fixes the `Subscription` streaming implementation, which
wraps over the upstream blocking NATS client API. Prior to this change,
calling `.next()`/`try_next()` on a `Subscription` stream would sit and
never yield a message, despite messages appearing on the target subject.

The issue here, was the prior version was inadvertently spawning a new
blocking task which was waiting on the next message on each call to
`.poll_next()`. The initial call would correctly immediately return a
`Poll::Pending` (i.e. the blocking task was still making progress), but
that future would never be polled again. Instead a *new* blocking task
would be spawned on the second `.poll_next()` call, and so on.

The fix here was to stash the running blocking task's `JoinHandle`, so
we can re-poll it on the next `.poll_next()` call, driving it to
completion.

Bonus Content
-------------

In addition to this fix the following improvements/tweaks are included:

- Update the `nats-subscribe` example to optionally provide a count of
  the maximum number of messages to process before closing the
  subscription.
- Add a `nats-publish` example to pair with `nats-subscribe` and behave
  a bit like `nats pub`.
- Add NATS subscription-specific metadata to each `Subscription` so that
  message-processing code can better instrument themselves. An example
  is provided in the `nats-subscibe` example, which will give us spans
  and timings of not only the total subscription receive time, but each
  message's processing timing.
- Update the instrumentation spans for the NATS client to be `debug`
  rather than the default of `info`. It's possible that `trace` might be
  the right level, although I think that could be reserved for lower
  level network/framing concerns. If this proves about right, then we
  should update the Posgres implementation's instrumentation levels as
  well.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>