feat(*): application create and list
Restores application creation and list functionality. Much of the UI for list is unimplemented - the cards are back, and the boxes are back, but the actual data inside them is not.

Otherwise, this creates a new component, assigns it to the `default` variant for the application `schema`, and allows you to list the components that use the `application` schema. It really shows off the delta between the old code base and the new code base.

Signed-off-by: Adam Jacob <adam@systeminit.com>