chore(web,sdf): restore application view and mock schematic panel
This PR brings back the 'application view', which is essentially the
application editor from the previous iteration. It also puts together a
simple mock panel for the schematic panel, and shows how to build simple
json based mock APIs in SDF. This is to support Alex in getting the
schematic code re-integrated, without having to worry about doing all
the work to get things all the way into the database.

A few things we know are buggy:

* The panel selector works, but doesn't show the correct selected panel
* The sample calls I show might not have the correct behavior, but get
  the gist across
* The attribute editor view does not match what we had shipped
  previously - its missing all sorts of context.

For Alex, you can should look in:

* app/web/organisims/PanelSchematic - the mock panel, with examples of
  calling the schematicservice (you could also not go all the way to SDF
  at all, and just have the service mock the return values - might be
  faster!)
* lib/sdf/src/server/service/schematic.rs - the backend of the
  schematicservice.

Signed-off-by: Adam Jacob <adam@systeminit.com>