chore(sdf,dal): support generating/using jwt secret keys in sdf
This is a first pass improvement on support JWT secret keys with respect
to the SDF (i.e. API) service.

There are 2 different "keys" at work when it comes to handling JWTs
which are used when users log into the platform:

1. An `RS256KeyPair`, which is a public/private keypair used to sign and
   verify that a token was issued by our service instance. This keypair
   is stored in the database so shared sdf instances all use the exact
   same keypair when confirming/issuing JWTs. In order to better protect
   the private part of the keypair at rest (in the database) it gets
   encrypted with...
2. A JWT secret key which is used to encrypt/decrypt the private `RS256`
   key in the database. The private key is encrypted on initial load and
   must be fetched and decrypted each time a JWT is issued (the
   verification of a JWT is done with the public key which is not
   encrypted by design).

Keypair Creation
---------------

The `RS256KeyPair` is generated and loaded into the database if not yet
present via the `install_new_jwt_key()` function which takes a
`JwtSecretKey`, necessary to encrypt the private key as noted above. At
the moment, `sdf` will check and possibly install a new keypair when it
performs its database migrations (via `Server::migrate_database()`). The
implication is that if `--migration-mode skip` is passed to `sdf`, then
no migrations are performed and this keypair check/installation is also
skipped by design. Later when multiple instances of `sdf` are running,
they likely run as a cluster in `--migration-mode skip` mode with an
ad-hoc run of `sdf --migration-mode runAndQuit` to perform initial
database setup and incremental migrations when deploying new versions of
the service.

JWT Secret Key Creation
-----------------------

In an effort to get us going and not messing with the CLI structure of
`sdf`, there is a new `--generate-jwt-secret-key <PATH>` option which
will generate a new suitable JWT secret key at the path provided--it
will then end the program and not run the service, mimicking a dedicated
subcommand which we may want to consider adding later.

JWT Secret Key Requirements for SDF
-----------------------------------

With this set of changes, `sdf` will now require a JWT secret key which
it will load when `sdf` boots to satisfy its signin/JWT-signing
responsibilities.

There are 4 strategies to tell `sdf` about where this key lives (first
match "wins"):

1. Provide a `--jwt-secret-key-path /some/path` option to `sdf`
2. Add a `jwt_secret_key_path = "/some/path"` in an `sdf.toml` (again,
   this can live in the current working directory--`./sdf.toml`, in
   `$HOME/.config/sdf/sdf.toml`, in `/usr/local/etc/sdf/sdf.toml`, or
   `/et/sdf/sdf.toml`)
3. Set an environment variable of
   `SI_SDF__JWT_SECRET_KEY_PATH=/some/path` for `sdf` to use
4. Use the current built-in default which right now is
   `/run/sdf/jwt_secret_key.bin`

The key location will be canonicalized (that is, determine the
"realpath" of the value) and confirmed to exist before it loads so a
missing key will result in `sdf` crashing on boot (by design).

**NOTE** at the moment `sdf` doesn't confirm that this key is sufficient
to decrypt any existing private key in the database, but we likely
should do this sanity check so `sdf` can crash early, rather than
bombing the request at the moment an end user attempts to login and be
issued a signed JWT.

JWT Secret Key Dev Hack
-----------------------

A final way that `sdf` can find an load its JWT secret key, is to run
`sdf` using `cargo run`. Currently, if `$CARGO_MANIFEST_DIR` is
detected, it will setup the default load location to be
`bin/sdf/src/dev.jwt_secret_key.bin` unless overridden by any of the
strategies above. This may be enough to satisfy "devmode" for the
moment, but it's still considered by this gent to be HACKASAURUS REX!

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>